requirements.txt
cloudml-hypertune
gcsfs
google-api-python-client
jinja2
joblib
kfp
pandas
pyyaml
scikit-learn
tensorflow>=0.14.0,<2.0.0
xgboost
